Fix issues with a nonstandard compiler and umask.
If the compiler is in a nonstandard location, libtool may choose the
incorrect mode.  This is described in libtool documentation:
     https://www.gnu.org/software/libtool/manual/html_node/Tags.html
The correct fix is to use --tag to explicitly set the mode.  See
https://sourceforge.net/p/openjade/patches/48/

Write permissions are removed when the package is built.  However, the
command chmod -w does not affect bits set in umask.  The correct usage
is to explicitly identify the users to which the command applies.  See
https://sourceforge.net/p/openjade/patches/49/