Add sile-0.10.9:
SILE is a typesetting system; its job is to produce beautiful printed
documents. Conceptually, SILE is similar to TeX - from which it borrows
some concepts and even syntax and algorithms - but the similarities end
there. Rather than being a derivative of the TeX family SILE is a new
typesetting and layout engine written from the ground up using modern
technologies and borrowing some ideas from graphical systems such as
InDesign.